[android] EditTextDialog moved to its own file, handles screen rotation
