+++ android/src/com/frostwire/android/gui/util/UIUtils.java
@@ -19,7 +19,6 @@ package com.frostwire.android.gui.util;
 
 import android.app.Activity;
 import android.app.AlertDialog;
-import android.app.Dialog;
 import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.DialogInterface;
@@ -27,21 +26,16 @@ import android.content.DialogInterface.OnClickListener;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.net.Uri;
-import android.os.Bundle;
 import android.os.Looper;
 import android.support.design.widget.Snackbar;
 import android.support.v4.content.FileProvider;
 import android.util.DisplayMetrics;
 import android.view.Gravity;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.view.Window;
 import android.view.inputmethod.InputMethodManager;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import com.andrew.apollo.utils.MusicUtils;
@@ -53,7 +47,7 @@ import com.frostwire.android.core.FileDescriptor;
 import com.frostwire.android.gui.Librarian;
 import com.frostwire.android.gui.activities.MainActivity;
 import com.frostwire.android.gui.services.Engine;
-import com.frostwire.android.gui.views.AbstractDialog;
+import com.frostwire.android.gui.views.EditTextDialog;
 import com.frostwire.util.Logger;
 import com.frostwire.util.MimeDetector;
 import com.frostwire.uxstats.UXAction;
@@ -176,10 +170,6 @@ public final class UIUtils {
         builder.create().show();
     }
 
-    public interface TextViewInputDialogCallback {
-        void onDialogSubmitted(String value, boolean cancelled);
-    }
-
     public static void showEditTextDialog(Activity activity,
                                           int messageStringId,
                                           int titleStringId,
@@ -187,7 +177,7 @@ public final class UIUtils {
                                           boolean cancelable,
                                           boolean multilineInput,
                                           String optionalEditTextValue,
-                                          final TextViewInputDialogCallback callback) {
+                                          final EditTextDialog.TextViewInputDialogCallback callback) {
         new EditTextDialog().
                 init(titleStringId,
                     messageStringId,
@@ -196,72 +186,6 @@ public final class UIUtils {
                     multilineInput,
                     optionalEditTextValue,
                     callback).show(activity.getFragmentManager());
-    }
-
-    public final static class EditTextDialog extends AbstractDialog {
-
-        private int titleStringId;
-        private int messageStringId;
-        private int positiveButtonStringId;
-        private boolean cancelable;
-        private boolean multilineInput;
-        private String optionalEditTextValue;
-        private TextViewInputDialogCallback callback;
-
-        public EditTextDialog() {
-            super(R.layout.dialog_default_input);
-        }
-
-        public EditTextDialog init(int titleStringId,
-                                   int messageStringId,
-                                   int positiveButtonStringId,
-                                   boolean cancelable,
-                                   boolean multilineInput,
-                                   String optionalEditTextValue,
-                                   final TextViewInputDialogCallback callback) {
-            this.titleStringId = titleStringId;
-            this.messageStringId = messageStringId;
-            this.positiveButtonStringId = positiveButtonStringId;
-            this.cancelable = cancelable;
-            this.multilineInput = multilineInput;
-            this.optionalEditTextValue = optionalEditTextValue;
-            this.callback = callback;
-            return this;
-        }
-
-        @Override
-        protected void initComponents(Dialog dlg, Bundle savedInstanceState) {
-            final TextView title = findView(dlg, R.id.dialog_default_input_title);
-            title.setText(titleStringId);
-            final EditText inputEditText = findView(dlg, R.id.dialog_default_input_text);
-            inputEditText.setHint(getString(messageStringId));
-            inputEditText.setMaxLines(!multilineInput ? 1 : 5);
-            if (optionalEditTextValue != null && optionalEditTextValue.length() > 0) {
-                inputEditText.setText(optionalEditTextValue);
-            }
-            final Button positiveButton = findView(dlg, R.id.dialog_default_input_button_yes);
-            positiveButton.setText(positiveButtonStringId);
-            final Button negativeButton = findView(dlg, R.id.dialog_default_input_button_no);
-            negativeButton.setText(R.string.cancel);
-            negativeButton.setVisibility(cancelable ? View.VISIBLE : View.GONE);
-            positiveButton.setOnClickListener(view -> {
-                if (inputEditText != null && callback != null) {
-                    try {
-                        callback.onDialogSubmitted(inputEditText.getText().toString(), false);
-                    } catch (Throwable t) {
-                        t.printStackTrace();
-                    }
-                }
-                dlg.dismiss();
-            });
-            negativeButton.setOnClickListener(view -> {
-                if (callback != null) {
-                    callback.onDialogSubmitted(null, true);
-                }
-                dlg.dismiss();
-            });
-        }
-
     }
 
     public static String getBytesInHuman(long size) {
